requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
email-validator==2.1.0
Pillow==10.2.0
pydub==0.25.1
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pytest==7.4.3
PyJWT==2.8.0
python-magic==0.4.27
